requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
Deprecated==1.2.13
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
PixivPy-Async==1.2.14
wrapt==1.15.0
yarl==1.8.2
